+++ packages/common/src/log/groupLog.ts
@@ -1,62 +1,50 @@
 // group log
 // e.g. glog("myGroup", yourdata...);
-import { LogLevel, logs } from "./log";
+import { LogLevel } from "./log";
 
-export const grouplogs = (
-  level: LogLevel,
-  isCollapsed: boolean,
-  groupName: string,
-  ...args: any[]
-) => {
-  if (isCollapsed) {
-    console.groupCollapsed(groupName);
-  } else {
-    console.group(groupName);
-  }
-  try {
-    logs(level, ...args);
-  } finally {
-    console.groupEnd();
-  }
+export const grouplogger = (
+  isCollapsed: boolean = false,
+  logLevel: LogLevel,
+): ((groupName: string, ...args: any[]) => void) => {
+  return (groupName: string, ...args: any[]) => {
+    let isStarted = false;
+    try {
+      if (isCollapsed) {
+        console.groupCollapsed?.(groupName);
+      } else {
+        console.group?.(groupName);
+      }
+      isStarted = true;
+
+      const logFn = console[logLevel] || console.log;
+      logFn(...args);
+    } catch (error) {
+      console.error(error);
+    } finally {
+      if (isStarted) {
+        console.groupEnd?.();
+      }
+    }
+  };
 };
 
 // group log =====
-export const grouplogInfo = (groupName: string, ...args: any[]) =>
-  grouplogs(LogLevel.info, false, groupName, ...args);
-
-export const grouplogDebug = (groupName: string, ...args: any[]) =>
-  grouplogs(LogLevel.debug, false, groupName, ...args);
-
-export const grouplogWarn = (groupName: string, ...args: any[]) =>
-  grouplogs(LogLevel.warn, false, groupName, ...args);
-
-export const grouplogError = (groupName: string, ...args: any[]) =>
-  grouplogs(LogLevel.error, false, groupName, ...args);
-
-export const grouplogTrace = (groupName: string, ...args: any[]) =>
-  grouplogs(LogLevel.trace, false, groupName, ...args);
+export const grouplogInfo = grouplogger(false, LogLevel.info);
+export const grouplogDebug = grouplogger(false, LogLevel.debug);
+export const grouplogWarn = grouplogger(false, LogLevel.warn);
+export const grouplogError = grouplogger(false, LogLevel.error);
+export const grouplogTrace = grouplogger(false, LogLevel.trace);
 
 // collapsed group log =====
-export const collapselogInfo = (groupName: string, ...args: any[]) =>
-  grouplogs(LogLevel.info, true, groupName, ...args);
-
-export const collapselogDebug = (groupName: string, ...args: any[]) =>
-  grouplogs(LogLevel.debug, true, groupName, ...args);
-
-export const collapselogWarn = (groupName: string, ...args: any[]) =>
-  grouplogs(LogLevel.warn, true, groupName, ...args);
-
-export const collapselogError = (groupName: string, ...args: any[]) =>
-  grouplogs(LogLevel.error, true, groupName, ...args);
-
-export const collapselogTrace = (groupName: string, ...args: any[]) =>
-  grouplogs(LogLevel.trace, true, groupName, ...args);
+export const collapselogInfo = grouplogger(true, LogLevel.info);
+export const collapselogDebug = grouplogger(true, LogLevel.debug);
+export const collapselogWarn = grouplogger(true, LogLevel.warn);
+export const collapselogError = grouplogger(true, LogLevel.error);
+export const collapselogTrace = grouplogger(true, LogLevel.trace);
 
 // alias =====
 export const glog = grouplogInfo;
-export const grouplog = grouplogInfo;
 export const clog = collapselogInfo;
-export const collapselog = collapselogInfo;
 
 export const glogInfo = grouplogInfo;
 export const glogDebug = grouplogDebug;