+++ packages/common/src/log/log.ts
@@ -7,31 +7,13 @@ export const LogLevel = {
 } as const;
 export type LogLevel = (typeof LogLevel)[keyof typeof LogLevel];
 
-export const logs = (level: LogLevel, ...args: any[]) => {
-  switch (level) {
-    case LogLevel.info:
-      logInfo(...args);
-      break;
-    case LogLevel.warn:
-      logWarn(...args);
-      break;
-    case LogLevel.error:
-      logError(...args);
-      break;
-    case LogLevel.debug:
-      logDebug(...args);
-      break;
-    case LogLevel.trace:
-      logTrace(...args);
-      break;
-  }
-};
+export type Log = (...args: any[]) => void;
 
-export const logInfo = (...args: any[]) => console.info(...args);
-export const logDebug = (...args: any[]) => console.debug(...args);
-export const logWarn = (...args: any[]) => console.warn(...args);
-export const logError = (...args: any[]) => console.error(...args);
-export const logTrace = (...args: any[]) => console.trace(...args);
+export const logInfo: Log = (...args: any[]) => console.info(...args);
+export const logDebug: Log = (...args: any[]) => console.debug(...args);
+export const logWarn: Log = (...args: any[]) => console.warn(...args);
+export const logError: Log = (...args: any[]) => console.error(...args);
+export const logTrace: Log = (...args: any[]) => console.trace(...args);
 
 // alias =====
 export const log = logInfo;